Add middleware functions for signup-form
Import dotenv into the app.js for hiding session secret.

Create the signup_get middleware, which renders the signup form view
with the title Sign Up.

Create the signup_post middleware, which sanitizes and validates the
data. If successful, the data then checks to see if there is another
user with the same username. If no user found, encrypt their password
and save the user to the database.